+++ pkg/repository/schema.sql
@@ -3,5 +3,5 @@ CREATE TABLE books (
     title varchar NOT NULL,
     description  text NOT NULL,
     author_name varchar NOT NULL,
-    price integer
+    price BIGINT
 );
